+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 Changelog
 =========
 
-##### Version 5.08.00 (WIP)
+##### Version 5.08.01 (WIP)
  * fix bugs
+
+##### Version 5.08.00 (2021-11-25)
  * fix broken testing suite
  * add page template types
+ * add archive frontend views
+ * add irradiador frontend views
 
 ##### Version 5.07.00 (2021-11-19)
  * fix missing assets install command execution during Ansible deploy